feat & refactor: 라즈베리파이 미소지자를 위한 알림 생성 및 밀집도 위험 알림 코드 분리
