+++ src/main/resources/static/situation.html
@@ -588,7 +588,7 @@
                     graphData.situationsDtos.forEach(function(i) {
                         cnt.push(i.cnt);
                     });
-                    var locName = new Array('A','B','C','D');
+                    var locName = new Array('가','나','다','라');
                     //그래프 x축 시간
                     function getCurrentWeek() {
                         var result = new Array();
